% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_qif.r
\name{parse_qif}
\alias{parse_qif}
\title{Parse a Qif file}
\usage{
parse_qif(source_file)
}
\arguments{
\item{source_file}{Path to the input file}
}
\value{
the Qif data in a data-frame
}
\description{
This function loads a Qif-formatted file, parses the file and displays the 
data in a structured way.
}
